requirements.txt
agate==1.9.1
altair==6.0.0
annotated-types==0.7.0
attrs==25.4.0
babel==2.18.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.22.5
dbt-common==1.37.2
dbt-core==1.11.2
dbt-duckdb==1.10.0
dbt-extractor==0.6.0
dbt-protos==1.0.427
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
duckdb==1.4.3
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.0
orderly-set==5.5.0
packaging==26.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
pillow==12.1.0
plotly==6.5.2
protobuf==6.33.5
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.1.0
sqlparse==0.5.4
streamlit==1.53.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
zipp==3.23.0
